refactor: moved contracts to the contracts folder out of the individual routes
If any of the Zod Schemas are shared with the frontend they cant be located in the server folder
