+++ src/app/conferences/create-conference.form.tsx
@@ -1,15 +1,12 @@
-"use client"
-
 import { zodResolver } from "@hookform/resolvers/zod"
 import currencyCodes from "currency-codes"
 import { Loader } from "lucide-react"
-import dynamic from "next/dynamic"
 import { type FC, useCallback, useMemo, useState } from "react"
 import { useForm } from "react-hook-form"
 import { NumericFormat } from "react-number-format"
-import rehypeSanitize from "rehype-sanitize"
 import { toast } from "sonner"
 
+import { MarkdownEditor } from "@/components/md-editor"
 import { Button } from "@/components/ui/button"
 import { DateTimePicker } from "@/components/ui/date-time-picker"
 import {
@@ -31,12 +28,8 @@ import {
 import {
   type CreateConferenceInput,
   CreateConferenceInputDto,
-} from "@/server/api/routers/conference/conference-create"
+} from "@/contracts/conference/conference"
 import { api } from "@/trpc/react"
-
-const MDEditor = dynamic(() => import("@uiw/react-md-editor"), {
-  ssr: false,
-})
 
 export type CreateConferenceProps = {
   close: () => void
@@ -119,19 +112,10 @@ export const CreateConferenceForm: FC<CreateConferenceProps> = ({
               <FormLabel>Description</FormLabel>
               <FormControl>
                 <div>
-                  <MDEditor
+                  <MarkdownEditor
+                    name={field.name}
                     value={field.value}
-                    onChange={(value) =>
-                      field.onChange({
-                        target: {
-                          value: value,
-                          name: field.name,
-                        },
-                      })
-                    }
-                    previewOptions={{
-                      rehypePlugins: [[rehypeSanitize]],
-                    }}
+                    onChange={field.onChange}
                   />
                 </div>
               </FormControl>
